+++ src/rpc/update_instance.rs
@@ -420,7 +420,12 @@ pub fn update_instance(
 
     if let Some(reason) = &req.target_state_reason {
         if reason != &instance.target_state_reason {
-            ops.assign(column_name!(Instance, target_state_reason), reason.clone())?;
+            if version_is_new_enough(
+                system_catalog_version,
+                &Instance::TARGET_STATE_CHANGE_TIME_AVAILABLE_SINCE,
+            )? {
+                ops.assign(column_name!(Instance, target_state_reason), reason.clone())?;
+            }
         }
     }